% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exo_planets_data}
\alias{exo_planets_data}
\title{Exoplanet dataset}
\usage{
exo_planets_data(exo_planets, col1_of_interest, col2_of_interest)
}
\arguments{
\item{exo_planets}{A dataframe containing all discovered exoplanets up until 2021. We only select those star systems with more than 1 planet, as they are not as dull and uninteresting.}

\item{col1_of_interest}{The user can select a column of their choice to check different things about the "interesting star systems. All columns are ok except "Distance' or 'Stellar Mass', as they are already provided}

\item{col2_of_interest}{The users wish is granted and a second column can be chosen to be investigated. All columns are ok except "Distance' or 'Stellar Mass', as they are already provided}
}
\value{
A simplified dataframe
}
\description{
Exoplanet dataset
}
\details{
This function returns a very simplified version of the original dataset, such that the person who is interested in studying the exoplanet dataset is not
overwhelmed and can digest the different features of the "interesting" planetary systems in great detail.
}
\examples{
exo_planets_data(exo_planets, num_stars, discovery_method)
}
